Does a report on a specific file if it is called with a path: ./cpclient.php report /var/log/syslog.
Using a trailing slash results in a nice stack trace, as well as using a path
which is not in the Catalog. I'll most likely use the difference of using a
trailing slash to display the directory entry as such (no trailing slash) vs.
displaying its contents. Remember, one name can exist as a directory and a file
in the Catalog at the same time!