Pretty obvious variation of catalog.php, the difference being that the files of a directory as well as the CatalogEntries are both fetched and then compared with each other by PHP. This however leaves a time window between getting the metadata of a file & actually storing it.
This brings down initial creation (empty database) down to 00:29:10 vs. 00:34:57
with catalog.php. This is still considerably slower than 00:05:43 with flat.php.
When executed against an existing index, flat.php is slower with 00:01:45 vs.
00:01:06 on catalog02.php and a tad bit faster than catalog.php with 00:01:53.

However, it should be kept in mind that we're talking about just creating an
empty catalog here. As soon as data gets backed up too, writing the index
will be dwarfed by the time it takes to copy the data.