A slightly modified implementation which uses byte precise offsets and lengths instead of blocks.
There is still a slight bug, 'tape' files can end up a few bytes larger as they
are supposed to be. It would be better to end up with a slightly smaller file.

Certain special cases should be considered as well, as to prevent a file from
having a few bytes on one tape and several Gigabytes on others, ie there has to
be some kind of algorithm that decides to use a new tape instead of using a few
kilobytes of another.