Added new net based client in src/include/Net, which is supposedly a 1:1 implementation of the local Client, with the difference being that Objects get serialized and sent from and to the server. This is of course inefficient as f…, but my plan is to get this working in the same way that the local client works and then start to optimize from there. Currently, this does not work, as it sends parent as an int instead of an instance of CatalogEntry, and it does not create any entries and therefore has to skip new directories.
Also, ModeNode will be changed in a way that the loop in ServerRunner ends and
ModeNode takes direct control over the server communication, as it will for one
be a binary interaction, and it will need to transfer large quantities of data.
It would be stupid to send a 10 GB iso file as string to onMessage.
I might decide to do a different implementation, but we will see.