Now files (not directories yet) are updated too. Skipping files if there is an UploadException.
I think I'll need some kind of restart mechanism in Protocol::sendFile. Doing it
from the Backup client with a loop like I did it in the Test would work, but
the Storage would use at worst five IDs before realizing that no file will be
sent.
If there is an additional onRestart, the Storage system could stick to whereever
it is writing too, starting anew.