Authentication for administrative client.
The whole authentication code is bull right now, there is no proper error
handling in the client. For now, I added a single OK which is not sent when
the client cannot authenticate, but I have to rewrite the whole ping-pong that
goes on between the server and the client on authenticating. Basically, it is
stupid to use Protocol /after/ authentication, which is a remainder of early
protocol tests with netcat.