Switched ssl-client/ssl-server (more precise, RunnerSSL) to use ProtocolReactive.
There is a hefty bug which I left in on purpose: the SSL-Server is not able to
discern the client sending commands. Which illustrates a major problem of the
event driven approach: context or lack of it. As I use RunnerSSL itself as the
ProtocolReactiveListener, there is next to no way for me to know which client
I am reacting to.
A common solution is to throw large amounts of parameters to the callback
around, preferably large objects, or use global scope. It even gets worse due to
the fact that humans tend to prefer anonymous functions - just look at common
JavaScript example and their usual mess of anonymous functions stacked into each
other.

As this is a learner project, I left the bug in on purpose, as I said.