I started to loathe the way I designed HubClientListener with the $name and $id parameters. However, I needed it for the forwarder. So I split it up into two variants, HubClientListener without parameters and HubClientNamedListener with the original approach. This way, I can boot out $name and $id from ProtocolReactive.
At the moment, HubClientListener and HubClientNamedListener are mutually
exclusive, a fact which is not caught as an error by StreamHub.

I keep in my mind of being able to use them both, ie that data read from a
client is, for instance, passed to ProtocolReactive and to RunnerSSL. This would
enable RunnerSSL to react to certain commands sent by the client.