Backup client now backups directories beyond and below the root directory. Ending the client does not work anymore, because it would end prematurely. Ah, the joys of async programming.
Added having the parent added to the constructor of CatalogEntries, which saves
a lot of hassle in the client, as the information of the current parent is not
available to the recursing function. Hmm, as I am writing this, I think of
another solution, but getting it from the server is faster than searching it
in a local array. Might change it anyway.

I think about abandoning the parent/child approach in d_catalog altogether. I
like to have that information, but it's more complicated. Just having the path
would be easier, although it would take more space in the database.

There is another reason to do this: having a "flat" catalog would enable me to
send a bunch of files & directories in one go, which would allow me to leverage
transactions on the server instead of lock db, write one entry, unlock db.