Added MockSender as well as MockSenderTest, courtesy of ChatGPT. Some observation:
-she made a mistake in line 22 of the test, expecting 88 bytes instead of 87.
-the test is more complete than the one of MockSender.
-she uses one instance of MockReceiver created in setUp, which is different from
 yesterday and could be seen as bad style because tests are not isolated
 anymore. However, since MockSender has no parameter, this is fine.
-she named the onSend/onCancel/onStart values differently
-she stuck to using true instead of TRUE

Yesterday's prompt was in German, today's in English.

I had quite the conversation with her about the quality of my code. She
correctly identified weak and strong points and missing parts.

On the other hand, I once asked a seasoned developer whether I should throw
this or that kind of exception. His answer was 'don't throw an exception at
all'. Yeah, why should I throw an exception if the user sends 'Second of
October 2023, the year of our Lord" if I can just convert it to 1970-01-01
instead!