Refactored ProtocolSync and ProtocolAsync to use SafeSender/SafeReceiver as a wrapper around whatever StreamSender/StreamReceiver is used by ProtocolSync and ProtocolAsync.
There is a side effect with SafeReceiver, which gets reused unintentionally.
$increment then is > 1 and the first control block is dropped. Did a quick fix
for this update and do a proper fix on the next.

Backing up does not work after the change - I need more unit tests and maybe
regression tests better than 'opening two terminals and have run code as root
on a production system' would be a reasonable investment too.