Kind of a philosophical question - should onSendEnd be called on cancelled transfers? For now, I say no, because otherwise, because this causes FileSender to close a resource which has already been closed.
This could of course be circumvented by more logic in FileSender, but I wanted
to keep implementations of StreamSender as stupid as possible.

I might rename onEnd into onSuccess to make it more clear that it is the
opposite to onCancel, and maybe have a generic onEnd which will be called either
way.