The last big obstacle in having a 'usable' system: aborted writes of new files would leave the catalog entry dangling. There will be a performance penalty for doing so at each new file, but I will revisit this later, as I need to do the same for directories as well.
This is commit #483 and it marks the point at which backup and restore
functionality works, from multiple clients in parallel. Not being able to
properly handle files that change during backup was the last real show stopper.

Of course, there are surely many bugs left, and more features and improvements,
and yes, I have to rebuild the reporting which I did not change yet after I
switched the catalog structure, but it is a good feeling nevertheless!