Changed n_version2basic in favor of more generic approach
Replaced n_version2basic with d_content, which has a column serial which is used
to generate a numeric id of a file stored in a 'basic' storage. This way, two
different storages have unique identifiers each. d_content will also be used
for better storages which are to be introduced in the future.