Transaction replaced by Mutex
Using transactions here may lead to the Database is locked error. While this is
a problem typical for SQLite and could be solved by using a 'proper' database,
I used Mutex here to make sure that both commands can only run as one. It does
not hurt if only the first command runs but not the second (ie, if the
server crashes in between).

Mutex is realized as an ad hoc solution here, however, I expect to be more use
cases for Mutexes, so a more centralized solution will be necessary.