Optimized StreamHub
The problem with StreamHub was that it included all streams in $write for
stream_select. stream_select passes on streams that have content to read or
which are writable - the latter being almost always the case. Now $write is
limited to streams to which needs to be written to, so stream_select will
actually block if there is nothing to read or write, as intended.
This brings down the time needed for a certain test run from about 00:01:41 to
00:01:04, which is considerable.