Fixed bug in FacadeProtocolListener and added NodeProtocolListener
Very clever to call the methods of FacadeProtocolListener on itself.
Fixed it and added a switch to NodeProtocolListener, which basically worked
out of the box.

There is, however, a bug that the client cannot read anymore after the server
has sent about 16 MB of data. This looks familiar, I think I had it in some
prototype, so I have to research it.