'Fixed' some issues with the client
As mentioned before, the server stopped after sending about 16 MB, which
was, however, just a number; the true reason being if GET CATALOG returns
a serialized value which is, if 5 bytes are added, an exact multiple of
1024.

I basically turned the protocol inside out, but was not able to find the
culprit. It also appears on the unchanged 'main' branch and I am quite
confident that it did not occur on PHP 7.4 or 7.2.