Fixed timeout bug
I always looked on the server, not the client: instruction was actually to
'read a book as long as more or zero pages are left". So the client tried to
read beyond the last page, but the book had no more pages, hence it took a while
before it got a 'value' ("") for the last page. As the protocol is strictly
question answer here, it was able to recover from that.