Going back to async protocol for backups
Going back to async allows to leverage the power of the Timeshare framework:
the directory iterator can run ahead while AsyncStream/AsyncProtocol are
reading or writing. There is a queue mechanism which stops the directory
iterator for some time if it is ahead too far.

Performance is still slow when uploading files, but shows a large increase
when scanning for changed data.

Backing up /etc with a max queue size of 1 - basically fallback to
synchronous protocol - takes about 13 seconds (5522 files, no upload).
A max queue size pushes it to 0.2 seconds.

And mind you, this is without real asynchronous programming.

For now, InEx list is hardcoded to include only /etc.

There are two bugs left:
- for some reason, DirectoryWalk sends directories two times on first backup
- I have not yet added the functionality to create the filesystem hierarchy,
  so backups aren't done properly when cpnc.php backup is used with any
  other path than the root directory.