Alternative directory iterator, as DirectoryWalk is bugged
Out of some reason, DirectoryWalk calls every file/directory twice, resulting
in a doubled number of files processed and regular crashes on the server side
due to Integrity Constraint Violations. I was not able to figure out the
problem, however, I know that I am quite fed up with the antics of either
FilesystemIterator (skips whole branches of the filesystem) or
RecursiveDirectoryIterator (slowest as far as I remember) or DirectoryIterator
(which is a pain due to getting the current file info as reference and has this
dual processing bug). Let's just stick to glob.
Also, SplFileInfo::isFile/isDir and is_file/is_dir have different opinions on
what is a file and what not.

RecurseDirectory has a more coarse grained runtime, as __tsLoop creates and
fill Files() for a whole directory instead of one file per __tsLoop call. It
would be possible to iterate over each array entry per __tsLoop call, should
the need arise.