Idle is an empty task which blocks for a selected time period
Without some kind of idling mechanism, the server will just push CPU usage to
100%, that is why PHP recommends to use stream_select before using fread,
fwrite or stream_socket_accept on nonblocking streams. However, doing so blocks
even if the server should run at full capacity.

Idle makes the server pause for some time, but Idle itself can be taken of the
stack using Scheduler::pause() when a new client connects, which is a simple
method to go to full throttle.

I thought about adding time based priorities to Schedule, to be more precise,
to timeshare. However, microtime or worse, hrtime is quite expensive and it
would need to be called on every loop.