Class to hold several small files and their metadata
There are many, many, many small files in a filesystem, and with small I'm
talking about a median of "below 4 kib".
For them, the existing method of file transfer is terribly inefficient, so I'll
add a method to transfer a whole group of files in one go.
The server then has to pick them apart and store them.