Unpack a class from binary data
It has to be said that serialize/unserialize is waaaaay faster than using
my binary writer which in the end uses pack/unpack, at the cost of being
larger.
I want, however, to evaluate FileGroup and I assume that it is a bad idea
to use serialize on binary data.