Files smaller than 1 Megabyte are collected and transferred in bulk
Currently, all files from a filegroup are stored at once, which of course
blocks the server and is a bad idea; however, as only 1 MB or 255 files
are written at once, this is acceptable for now.

serialize is actually way faster that using my own toBinary/fromBinary approach;
in general, pack/unpack is a resource hog.
As the binary file data is stored as binary in a serialize string, additional
overhead is negligible.