cpnc.php restore blocks indefinitely if server terminates connection
cpnc backup still does not add parent directories when not starting at the
root directory ie if /var/ is backed up, /var/ itself is not created, resulting
so far in an error if /var/ was looked up from the catalog by restore - and
since restore still uses a blocking stream, it blocks eternally.
For now, server catches exception and sends a message, which restore can't
handle.