Last FileGroup was not sent.
Protol sent "DONE" as soon as onEnd() was called by the iterator, but without
sending any remaining filegroup, which is only done as a reaction to getting
a serialized instance of File from the server. So either end at onEnd or
onSerialized after everything has been uploaded.

TaskSingleStorage will finish on the server until after the BA client has
disconnected, which is a question of philosophy I think. I'd prefer for the
BA client to finish only if it knew that everything is accounted for; I will
look into that at some point.