Changed handling of idle process
Idle would be restarted after the client disconnected, but this meant that
TaskSingleStorage would be slowed down as well, although not considerably.

However, more tasks can be expected in the future, therefore I jerryrigged some
distinction that only two processes - Server\TaskServer and Server\Input -
are allowed to be slowed down by Idle. cpadmc connections could be slowed down
up to a point, but I need to rethink the details.

I thought about implementing a native Idle solution to Timeshare, but I found
the cost of hrtime/microtime prohibitive. However, I may revisit it using a
similar approach to the one used here, as pause/resume is elegant.