Added transaction to speed up BasicStorage::storeSingle()
Putting the necessary database interactions into a transaction sped up things
quite considerably; for /usr/share/help - 84k files - I've got 153 to 177
seconds with 0.28 to 0.29% spent for getting the storage id and 0.67 % to
0.77 % for finalizining (telling d_content/d_version that the file has indeed
been stored).
Without transaction, it is 240 to 242 seconds as well as 16.15 to 16.45 %
and 32.82% to 32.87 % spent for the database operations.

This translates to 'the time spent on database operations is now negligible'.