Added interface BinaryPersistable, used it for File
Classes are BinaryPersistable if they can be turned into a binary string and
loaded from a binary string. This will be safer than serialize/unserialize,
because the class name will neet to be known prior to using Klass::fromString.
However, ProtocolAsync is insecure in it's current iteration, as it will
happily instantiate whatever class has an fromString method. While this greatly
reduces the potential classes an attacker has at his disposal, it still needs to
be thought of.