Some additional types.
I also added some remarks about the relationship $streamReceiver and
$fileReceiver, which is not as obvious.

Also, I added a "pseudo type cast" for $streamReceiver to StringReceiver,
so that $this->streamReceiver is not called on 'good luck'. This would
result in a more evident Type Error as opposed to no such method or what
would be the error if called on the wrong instance.

However, Psalm does not like it if more general types are used. To be
honest, yes, it could be done by mistake, but it is OOP in the end...