Added some psalm findings & removed obsolete code.
This is a nice one - when starting with the project, I started with a local
system. CPAdm was an 'administrative' client working directly on the database.
It is still used in tests due to its handle command function.
While deprecated at first glance, it is actually good to have around - I could
increase test coverage by using it to test server functions without a running
server.
For Node command, it should be possible to isolate the node protocols.