Some level 1 findings of Psalm corrected
However, I decided to stay with level 4 for the moment. Beyond level 4, most
of the problems will be stuff like MixedArrayAccess and so on, which is caused
by the fact that PHPs arrays are not type safe; this is further complicated by
the fact that PDOStatement returns mixed as well and this then propagates its
way all to the top. So I basically would have to write a type safe version of
PDO/PDOStatement and Arrays.