Links are handled as usual (stored in the data part of a file on the server)
I am currently pondering about the whole architecture. So far, a block of 8kb
was added to every file, which was cut off by the server before restoring a
file, as the client knew what file it requested.
But now I am sending the header along with every file anyway, so I might as well
start the file directly after the header (which would complicate stream
reading).