Rewrote CatFileDiff to be more generic
Instead of deleted, changed and new, which are mostly relevant from the backup
perspective, it has now serverOnly, clientOnly as well as different; different
is split up between files and catalog entries. When backing up, one would want
to get the differing files (to send them to the server), when restoring, one
would want the differing catalog entries (to know which version of a file to
request).