Added onStreamStart and on StreamEnd to ProtocolAsyncListener
File Streams are implemented in a way that ProtocolAsync relinquishes
control to SafeSender and the file stream it was given to, and returns to
normal proceedings once the file has been received. But the protocol listener
was entirely left out so far, which I have changed by adding onStreamStart
and onStreamEnd.

Also, I fixed a bug in the test for files < 1024 bytes which did not transfer
the file completely. SafeSender transfers a file by sending at least 1kb
of header, then 1kb of payload and another 1kb for a control block which
tells FileSender that the file was transferred allright, but I only ran read
once, so just an empty file was created.