Greatly enhanced restore speed, restore is about as fast as backup.
Implementation is not very good and thus not final. Basically, RestoreTask runs
in parallel to AsyncStream and controls the restore. RestoreProtocolListener
is in itself not able to pause for some time, as it has no way to resume
operations. RestoreTask, on the other hand, does.

What will be changed are the public properties in RestoreProtocolListener, which
are accessed directly by RestoreTask.